+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 0.9.0
+
+- Support reading excerpts from `*.excerpt.yaml` files when the new `--yaml` flag is used.
+- Support `plaster` attributes. For legacy excerpts, only `plaster="none"` is recognized.
+
 ## 0.8.1
 
 - Support `code-excerpt` instructions as the first line